workspaces/1/projects/amazon--chronos-t5-small/requirements.txt
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
chronos-forecasting>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
